Updates iCloud client
