+++ app/clients/icloud/init.js
@@ -1,9 +1,17 @@
-const monitorMacServerStats = require('./util/monitorMacServerStats');
-const resync = require('./util/resyncRecentlySynced');
+const monitorMacServerStats = require("./util/monitorMacServerStats");
+const resync = require("./util/resyncRecentlySynced");
+const initialTransfer = require("./sync/initialTransfer");
+const database = require("./database");
 
 module.exports = async () => {
 
+  await database.iterate(async (blogID, account) => {
+    if (account.transferringToiCloud) {
+      await initialTransfer(blogID);
+    }
+  });
+
   resync();
-  
+
   monitorMacServerStats();
 };
