+++ app/clients/icloud/util/resyncRecentlySynced.js
@@ -3,7 +3,7 @@ const RESYNC_WINDOW = 1000 * 60 * 10; // 10 minutes
 const clfdate = require("helper/clfdate");
 
 const database = require("../database");
-const resync = require("../sync/from_iCloud");
+const resync = require("../sync/fromiCloud");
 const Blog = require("models/blog");
 
 const getLastSyncDateStamp = (blogID) => {
@@ -25,15 +25,20 @@ module.exports = async () => {
   console.log(clfdate(), "Resyncing recently synced blogs");
 
   await database.iterate(async (blogID, account) => {
+  
+    if (!account.setupComplete) {
+      console.log(
+        clfdate(),
+        "Account setup not complete, skipping resync: ",
+        blogID
+      );
+      return;
+    }
+
     const lastSync = await getLastSyncDateStamp(blogID);
 
     if (!lastSync) {
       console.log(clfdate(), "No last sync date found for blogID: ", blogID);
-      return;
-    }
-
-    if (account && account.transferringToiCloud)  {
-      console.log(clfdate(), "Blog is transferring to iCloud, skipping resync: ", blogID);
       return;
     }
