+++ app/clients/icloud/views/index.html
@@ -22,14 +22,14 @@
 {{> status-line}}
 
 {{#sharingLink}}
-{{^setupComplete}}
+{{^acceptedSharingLink}}
 {{^error}}
 <div class="line">
   <span class="label">Setup</span>
   <span class="center">Setting up folder...</span>
 </div>
 {{/error}}
-{{/setupComplete}}
+{{/acceptedSharingLink}}
 {{/sharingLink}}
